server/requirements.txt
fastapi==0.95.2
pydantic==1.10.7
uvicorn==0.22.0
sqlalchemy==2.0.13
alembic==1.11.0
python-dotenv==1.0.0
psycopg2==2.9.6
typer[all]==0.9.0